requirements.txt
flask
qrcode
pillow
pyzbar
